third_party/blink/web_tests/platform/generic/external/wpt/css/css-grid/parsing/grid-shorthand-serialization-expected.txt

This is a testharness.js-based test.
PASS Serialization without grid-template-areas
PASS Serialization with grid-template-areas
FAIL Serialization with grid-auto-flow assert_equals: grid shorthand must not be serialized when it doesn't roundtrip expected "" but got "30px 40px / auto-flow auto"
Harness: the test ran to completion.
